Extract fields.js file
